Adds setup.py configuration for installation via pip etc
This enables including the modules in this repository to be easily
included in conda environment specifications or to be installed in
a virtual environment via `pip install .`, or by installing from
the Git repository.

In this version it enumerates as modules to be installed all .py
files starting with a capital letter. Scripts etc in all-lowercase
files are therefore not installed, but this could be added where
useful. Also, it is probably better then to create a proper package
structure.